requirements.txt
pandas
discord.py
audioop-lts
python-dotenv